move img to its own subfolder
