+++ cmake/Dependencies.cmake
@@ -5,6 +5,7 @@ if (CMAKE_CROSSCOMPILING)
 endif()
 
 if (BUILD_THERION OR BUILD_LOCH)
+    add_subdirectory(extern/img)
     find_package(fmt REQUIRED)
 endif()
