+++ extern/img/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_library(img img.c)
+target_compile_definitions(img PUBLIC IMG_API_VERSION=1)
+target_include_directories(img SYSTEM PUBLIC  ${CMAKE_CURRENT_SOURCE_DIR}/include
+                                      PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/include/img)
+target_link_libraries(img PRIVATE disable-warnings)
