+++ loch/CMakeLists.txt
@@ -4,11 +4,6 @@ endif()
 
 set(LOCH_DEFINITIONS _GNU_SOURCE LOCH "LX${THPLATFORM}" GL_SILENCE_DEPRECATION TBB_SUPPRESS_DEPRECATED_MESSAGES=1)
 
-# extern library img
-add_library(img img.h img.c)
-target_compile_definitions(img PUBLIC IMG_API_VERSION=1)
-target_link_libraries(img PRIVATE disable-warnings)
-
 # library common with therion
 add_library(common-utils STATIC icase.h icase.cxx lxFile.h lxFile.cxx lxMath.h lxMath.cxx)
 target_compile_definitions(common-utils PUBLIC $<$<BOOL:${WIN32}>:WIN32_LEAN_AND_MEAN> PRIVATE ${LOCH_DEFINITIONS})
