--- loch/img.c
@@ -1,1 +1,0 @@
-#include "../extern/img.c"
