--- loch/img.h
@@ -1,1 +1,0 @@
-#include "../extern/img.h"
